# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/advanced/error-correction/codes": {
      "get": {
        "operationId": "AdvancedFeaturesController_getQECCodes",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Advanced"
        ]
      }
    },
    "/api/v1/advanced/error-correction/{codeId}/encode": {
      "post": {
        "operationId": "AdvancedFeaturesController_encodeCircuit",
        "parameters": [
          {
            "name": "codeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Encode a logical state with the chosen code.",
        "tags": [
          "Advanced"
        ]
      }
    },
    "/api/v1/advanced/error-correction/syndrome": {
      "post": {
        "operationId": "AdvancedFeaturesController_measureSyndrome",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Measure the syndrome of an encoded logical state.",
        "tags": [
          "Advanced"
        ]
      }
    },
    "/api/v1/advanced/noise/channels": {
      "get": {
        "operationId": "AdvancedFeaturesController_getNoiseChannels",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Advanced"
        ]
      }
    },
    "/api/v1/advanced/noise/apply": {
      "post": {
        "operationId": "AdvancedFeaturesController_applyNoise",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Validate a set of noise channels against the model.",
        "tags": [
          "Advanced"
        ]
      }
    },
    "/api/v1/advanced/noise/characterize": {
      "post": {
        "operationId": "AdvancedFeaturesController_characterizeNoise",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Characterize a device from one of the built-in noise models.",
        "tags": [
          "Advanced"
        ]
      }
    },
    "/api/v1/advanced/ml/vqe/ansatz": {
      "get": {
        "operationId": "AdvancedFeaturesController_getVQEAnsatzTypes",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Advanced"
        ]
      }
    },
    "/api/v1/advanced/ml/vqe/run": {
      "post": {
        "operationId": "AdvancedFeaturesController_runVQE",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Run a real VQE optimization.",
        "tags": [
          "Advanced"
        ]
      }
    },
    "/api/v1/advanced/ml/kernel/matrix": {
      "post": {
        "operationId": "AdvancedFeaturesController_getKernelMatrix",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Compute a real quantum kernel matrix for the given data.",
        "tags": [
          "Advanced"
        ]
      }
    },
...
    "/io/import": {
      "post": {
        "operationId": "IOController_importCircuit",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Import circuit from external format.",
        "tags": [
          "IO"
        ]
      }
    },
    "/io/export": {
      "post": {
        "operationId": "IOController_exportCircuit",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Export circuit to external format.",
        "tags": [
          "IO"
        ]
      }
    },
    "/io/convert": {
      "post": {
        "operationId": "IOController_convertCircuit",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConvertDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Convert between formats.",
        "tags": [
          "IO"
        ]
      }
    },
    "/io/validate": {
      "post": {
        "operationId": "IOController_validateCircuit",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Validate circuit data.",
        "tags": [
          "IO"
        ]
      }
    },
...
    "/io/formats/{name}": {
      "get": {
        "operationId": "IOController_getFormatInfo",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get format adapter info.",
        "tags": [
          "IO"
        ]
      }
    },
...
  },
  "info": {
    "title": "casimirQ API",
    "description": "REST API for the casimirQ quantum circuit simulation platform: authentication, circuits, simulation, asynchronous jobs, algorithms, visualizations, and advanced features (error correction, noise, quantum ML).",
    "version": "1.0.0",
    "contact": {},
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Docker (nginx single-origin)"
    },
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local backend"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ImportDto": {
        "type": "object",
        "properties": {}
      },
      "ExportDto": {
        "type": "object",
        "properties": {}
      },
      "ConvertDto": {
        "type": "object",
        "properties": {}
      },
      "ValidateDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}